+++ code/config/ConfigurationsReader.java
@@ -205,7 +205,7 @@ public class ConfigurationsReader {
 					Configurations.setUseSkywalking(Boolean.parseBoolean(arr[1].trim()));
 					break;
 				case "mergeWeight":
-					Configurations.setMergeWeight(Double.parseDouble(arr[1].trim()));
+					Configurations.setMergeWeight(Integer.parseInt(arr[1].trim()));
 					break;
 				case "quantileNum":
 					Configurations.setQuantileNum(Integer.parseInt(arr[1].trim()));
